requirements.txt
numpy~=1.21.5
torch~=1.8.1
torchvision~=0.12.0
blobfile~=1.3.1
lmdb~=1.3.0
pillow~=9.1.1
setuptools~=59.5.0
requests~=2.26.0
tensorflow~=2.9.1
tensorflow-gpu~=2.5.0
scipy~=1.7.1
tqdm~=4.62.3
mpi4py~=3.1.3